Function ncollide_queries::geometry::algorithms::gjk::project_origin

pub fn project_origin<N, P, V, M, S, G: ?Sized>(m: &M, shape: &G, simplex: &mut S) -> Option<P> where N: Scalar, P: Point<N, V>, V: Vect<N>, S: Simplex<N, P>, G: SupportMap<P, V, M>

Projects the origin on a shape unsing the GJK algorithm.

Arguments:

shape - the shape to project the origin on

simplex - the simplex to be used by the GJK algorithm. It must be already initialized with at least one point on the shape boundary.