Struct ncollide_queries::geometry::algorithms::johnson_simplex::RecursionTemplate

pub struct RecursionTemplate {
    // some fields omitted
}

Set of indices to explain to the JohnsonSimplex how to do its work. Building this is very time consuming, and thus should be shared between all instances of the Johnson simplex.

Methods

impl RecursionTemplate

fn new(dim: usize) -> Arc<Vec<RecursionTemplate>>

Creates a new set of Recursion simplex sharable between any Johnson simplex having a dimension inferior or equal to dim.