Trait ncollide_queries::point::PointQuery

pub trait PointQuery<N, P, M: Transform<P>>: LocalPointQuery<N, P> {
    fn project_point_with_transform(&self, m: &M, pt: &P, solid: bool) -> P { ... }
    fn distance_to_point_with_transform(&self, m: &M, pt: &P) -> N { ... }
    fn contains_point_with_transform(&self, m: &M, pt: &P) -> bool { ... }
}

Trait of objects that can be transformed and tested for point inclusion and projection.

Provided Methods

fn project_point_with_transform(&self, m: &M, pt: &P, solid: bool) -> P

Projects a point on self transformed by m.

fn distance_to_point_with_transform(&self, m: &M, pt: &P) -> N

Computes the minimal distance between a point and self transformed by m.

fn contains_point_with_transform(&self, m: &M, pt: &P) -> bool

Tests if the given point is inside of self transformed by m.