Trait ncollide_queries::ray::LocalRayCast

pub trait LocalRayCast<N, P, V> {
    fn toi_and_normal_with_ray(&self, ray: &Ray<P, V>, solid: bool) -> Option<RayIntersection<N, V>>;

    fn toi_with_ray(&self, ray: &Ray<P, V>, solid: bool) -> Option<N> { ... }
    fn toi_and_normal_and_uv_with_ray(&self, ray: &Ray<P, V>, solid: bool) -> Option<RayIntersection<N, V>> { ... }
    fn intersects_ray(&self, ray: &Ray<P, V>) -> bool { ... }
}

Traits of objects which can be tested for intersection with a ray.

Required Methods

fn toi_and_normal_with_ray(&self, ray: &Ray<P, V>, solid: bool) -> Option<RayIntersection<N, V>>

Computes the intersection point between this shape and a ray.

Provided Methods

fn toi_with_ray(&self, ray: &Ray<P, V>, solid: bool) -> Option<N>

Computes the time of impact between this shape and a ray

fn toi_and_normal_and_uv_with_ray(&self, ray: &Ray<P, V>, solid: bool) -> Option<RayIntersection<N, V>>

Computes the intersection point and normal between this shape and a ray.

fn intersects_ray(&self, ray: &Ray<P, V>) -> bool

Tests whether a ray intersects this shape.