Trait ncollide_queries::geometry::algorithms::simplex::Simplex

pub trait Simplex<N, P> {
    fn reset(&mut self, P);
    fn add_point(&mut self, P);
    fn project_origin_and_reduce(&mut self) -> P;
    fn project_origin(&mut self) -> P;
    fn contains_point(&self, &P) -> bool;
    fn dimension(&self) -> usize;
    fn max_sq_len(&self) -> N;
    fn modify_pnts(&mut self, f: &Fn(&mut P));
}

Trait of a simplex usable by the GJK algorithm.

Required Methods

fn reset(&mut self, P)

Replace the point of the simplex by a single one. The simplex is reduced to be 0-dimensional.

fn add_point(&mut self, P)

Adds a point to the simplex.

fn project_origin_and_reduce(&mut self) -> P

Project the origin on the simplex and remove any sub-simplex which does not contain the projection.

fn project_origin(&mut self) -> P

Projection the origin on the simplex. The simplex itself is unchanged, although it is mutable for optimization purpose.

fn contains_point(&self, &P) -> bool

Checks whether a given point is already part of the simplex points.

fn dimension(&self) -> usize

Dimension of the simplex. A simplex with n must be a n - 1-dimensional simplex.

fn max_sq_len(&self) -> N

The maximum among the simplex point squared lengths.

fn modify_pnts(&mut self, f: &Fn(&mut P))

Modifies the points contained by this simplex.