Module ncollide_queries::ray

Ray-casting related definitions and implementations.

Structs

Ray
A Ray.

RayInterferencesCollector
Bounding Volume Tree visitor collecting interferences with a given ray.

RayIntersection
Structure containing the result of a successful ray cast.

RayIntersectionCostFn
A search thet selects the objects that has the smallest time of impact with a given ray.

Traits

LocalRayCast
Traits of objects which can be tested for intersection with a ray.

RayCast
Traits of objects which can be transformed and tested for intersection with a ray.

Functions

ball_toi_with_ray
Computes the time of impact of a ray on a ball.

implicit_toi_and_normal_with_ray
Cast a ray on a shape using the GJK algorithm.

plane_toi_with_ray
Computes the toi of a ray with a plane described by its center and normal.

triangle_ray_intersection
Computes the intersection between a triangle and a ray.

Type Definitions

Ray2
A 2D ray.

Ray3
A 3D ray.

RayIntersection2
A 2D ray intersection.

RayIntersection3
A 3D ray intersection.